+++ Polly.Contrib.AzureFunctions/FooFragileFunctionConsumingBreaker_ConsistencyPriority.cs
@@ -11,36 +11,36 @@ using Polly.Contrib.AzureFunctions.CircuitBreaker;
 
 namespace Polly.Contrib.AzureFunctions
 {
-    public class FooFragileFunctionConsumingBreaker_FidelityPriority
+    public class FooFragileFunctionConsumingBreaker_ConsistencyPriority
     {
         // Uniquely identifies the circuit-breaker instance guarding this operation.
-        private const string CircuitBreakerId = nameof(FooFragileFunctionConsumingBreaker_FidelityPriority);
+        private const string CircuitBreakerId = nameof(FooFragileFunctionConsumingBreaker_ConsistencyPriority);
 
         // Used by this demonstration code to generate random failures of the simulated work.
         private static readonly Random Rand = new Random();
 
         private readonly IDurableCircuitBreakerOrchestrator durableCircuitBreakerOrchestrator;
 
-        public FooFragileFunctionConsumingBreaker_FidelityPriority(IDurableCircuitBreakerOrchestrator durableCircuitBreakerOrchestrator)
+        public FooFragileFunctionConsumingBreaker_ConsistencyPriority(IDurableCircuitBreakerOrchestrator durableCircuitBreakerOrchestrator)
         {
             this.durableCircuitBreakerOrchestrator = durableCircuitBreakerOrchestrator;
         }
 
 
-        [FunctionName("FooFragileFunctionConsumingBreaker_FidelityPriority")]
+        [FunctionName("FooFragileFunctionConsumingBreaker_ConsistencyPriority")]
         public async Task<IActionResult> Run(
             [HttpTrigger(AuthorizationLevel.Function, "get", "post", Route = null)] HttpRequestMessage req,
             ILogger log,
             [OrchestrationClient]IDurableOrchestrationClient orchestrationClient
         )
         {
-            // In the _FidelityPriority example, we await hearing from the circuit-breaker whether execution is permitted.
+            // In the _ConsistencyPriority example, we await hearing from the circuit-breaker whether execution is permitted.
             // This makes operations entirely faithful to the state of the breaker at any time,
             // and allows us to restrict executions in the half-open state to a limited number of trial executions.
 
-            if (!await durableCircuitBreakerOrchestrator.IsExecutionPermittedByBreaker_FidelityPriority(orchestrationClient, CircuitBreakerId, log))
+            if (!await durableCircuitBreakerOrchestrator.IsExecutionPermittedByBreaker_ConsistencyPriority(orchestrationClient, CircuitBreakerId, log))
             {
-                log.LogError($"{nameof(FooFragileFunctionConsumingBreaker_FidelityPriority)}: Service unavailable.");
+                log.LogError($"{nameof(FooFragileFunctionConsumingBreaker_ConsistencyPriority)}: Service unavailable.");
 
                 return new StatusCodeResult((int)HttpStatusCode.ServiceUnavailable);
             }
@@ -57,7 +57,7 @@ namespace Polly.Contrib.AzureFunctions
             {
                 await durableCircuitBreakerOrchestrator.RecordFailureForBreaker(orchestrationClient, CircuitBreakerId, log);
 
-                log.LogError(exception, $"{nameof(FooFragileFunctionConsumingBreaker_FidelityPriority)}: Exception: {exception.Message}");
+                log.LogError(exception, $"{nameof(FooFragileFunctionConsumingBreaker_ConsistencyPriority)}: Exception: {exception.Message}");
 
                 return new InternalServerErrorResult();
             }